Fix bug in pre-compiled .so library
chicken-install compiles the library to a .so file using csc. This
pre-compiled library is then loaded by csi. The library was using
cond-expand to figure out whether csi or csc is being used; the
cond-expand is expanded by csc and is not evalutated by the object
code, so loading it into csi would give the wrong impression that
we're in compiled code.

Fix by replacing `cond-expand` with runtime calls to `feature?`.
Thanks to Felix Winkelmann for solving this.